+++ src/analysis/rank/DividendRank.ts
@@ -1,0 +1,9 @@
+import { Rank } from "./rank";
+
+export class DividendRank extends Rank {
+    protected async internalSort() {
+  
+    }
+}
+  
+
