+++ src/analysis/rank/PERank.ts
@@ -1,0 +1,9 @@
+import { Rank } from "./rank";
+
+export class PERank extends Rank {
+    protected async internalSort() {
+  
+    }
+  }
+  
+
