+++ src/analysis/rank/ROERank.ts
@@ -1,0 +1,8 @@
+import { Rank } from "./rank";
+
+export class ROERank extends Rank {
+    protected async internalSort() {
+  
+    }
+}
+
